+++ src/protocolgame.h
@@ -169,7 +169,7 @@ private:
 	void sendFYIBox(const std::string& message);
 
 	void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
-	void sendMagicEffect(const Position& pos, uint8_t type);
+	void sendMagicEffect(const Position& pos, uint16_t type);
 	void sendCreatureHealth(const Creature* creature);
 	void sendSkills();
 	void sendPing();
